WIP rip off swapPluginsAndMasterVersion: SSS TTT
Note swapPluginsAndMasterVersion would be called *after* set_local_save
which sets the Bash title - so if the esms were swapped the title would
not reflect that.

I ended up introducing try_set_version to decouple save and modInfos -
not sure if voAvailable may contain voCurrent but let's not pursue this
further.