Drop the error checks TTT
Trust _voAvailabe, voCurrent are up to date

FFF check if move_to exists in our cache - this also catches the case
move_to be present and ghosted which the exists check would miss - we
have to be careful and check corrupted too (and trust those are up to
date too). API limits here.